Move form logic into parent component
